feat: Load & save conversations depending on UUID
I've integrated the FingerprintJS
library to create an UUID string.

This string will be used to identify
the user on the API endpoint.

If the UUID doesn't exist yet, a new
conversation will be created with two
default greeting messages from Parola.

Consider: There is still a bug with
the initiation of a new conversation.

The try and catch block doesn't work
as intended. If you visit Parola
for the first time, it will yield
an error. But after refreshing, it
will display the correct data from
the DB with the default greeting
messages.