fix: Added zero-trust encryption (not 100 % done)
I use the FingerprintJS UUID string
to encrypt and decrypt all user
related data on the DB.

The only thing that's currently
not very secure is that the
initial vector for the hashing
process of the UUID is static
and not randomly generated.

I decided to move on to
server-side fingerprinting
and implement the FingerprintJS Pro
service and maybe find some random
generated data which I can use
as an initial vector.

My goal is not to create the
initial vector on the server
since I would need to save this as
well on the DB. The logic there
doesn't work.

Right now I will close this branch
and start working on the FingerprintJS
Pro part of the project.