feat: Added FingerprintJS Pro
I purchased FingerprintJS Pro and
created a token to use their web
service to identify the users with
a way higher accuracy as with
the free version.

Since the UserID element of the
pro version has a smaller char length
I needed to hash it so that it fits
the pre-defined key length of
the NodeJS Crypto module.