feat: Added DialogFlow to the back-end
DialogFlow has now finally been added
to the back-end. It also works with
the encryption module.

I needed to change the logic of the
front-end a little (but not in a
exquisite way) => TODO for later.

Now I will refactor everything and
embed NextJS for the client.

But for now, I will work on the
DialogFlow side of things (creating
a Swiss-German agent in GCP).